pyproject.toml
[project]
name = "mcpt"
version = "0.1.0"
description = "CLI for discovering and running MCP Tool Shop tools."
requires-python = ">=3.10"
dependencies = [
    "typer>=0.12",
    "pyyaml>=6.0",
    "rich>=13.0",
    "platformdirs>=4.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
]

[project.scripts]
mcpt = "mcpt.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["mcpt"]
